+++ Projects/2D_Roguelike/MoveState.cpp
@@ -10,7 +10,7 @@ void MoveState::Enter(Player& agent)
 void MoveState::Update(Player& agent, float deltaTime)
 {
 	float tilesize = TileMap::Get().GetTileSize();
-	if (!agent.path.empty() && agent.currentPath < agent.path.size())
+	if (!agent.path.empty() && static_cast<size_t>(agent.currentPath) < agent.path.size())
 		agent.destination = 
 		{agent.path[agent.currentPath].x*tilesize ,agent.path[agent.currentPath].y*tilesize + tilesize };
